+++ py-polars/tests/unit/namespaces/test_list.py
@@ -79,6 +79,12 @@ def test_contains() -> None:
     assert_series_equal(out, expected)
 
 
+def test_list_contains_invalid_datatype() -> None:
+    df = pl.DataFrame({"a": [[1, 2], [3, 4]]}, schema={"a": pl.Array(pl.Int8, width=2)})
+    with pytest.raises(pl.SchemaError, match="invalid series dtype: expected `List`"):
+        df.select(pl.col("a").list.contains(2))
+
+
 def test_list_concat() -> None:
     df = pl.DataFrame({"a": [[1, 2], [1], [1, 2, 3]]})
